requirements.txt
Django==1.11.11
django-leaflet==0.23.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
gunicorn==19.8.1
psycopg2==2.6
PyJWT==1.5.3
python-decouple==3.1
pytz==2018.3
